Add example of SOLT calibration and application
This is a simple example of using SOLT calibration to solve for the
error terms in a simulated flawed VNA, then using the resulting error
terms to correct the raw measurements of a simulated DUT.  The example
closely follows vnacal-SOLT-example.c from libvna.

Renamed and fixed the other example programs for consistency.