Calset.add(solver, name) -> Solver.add_to_calset(name)
In the C version of libvna, vnacal_add_calibration is an operation of
vnacal_t because it fits into a family of related functions including:
vnacal_find_calibration and vnacal_delete_calibration.

In the python version, however, the other functions are dispersed as
Calset.index and _CalHelper.__delitem__, leaving the "add" function alone.
It makes sense then to move the add function into Solver, where it can
take just one argument instead of two.