Simplify Calset.properties and Calibration.properties usage
Before, reading from the properties attributes returned a deep copy of
the properties, mirroring the C version.  To make a change, you could
edit your copy, but then you had to assign the root back to the Calset
or Calibration properties attribute to make the change persistent.
This was error prone because it was too easy for a user to make the
mistake of editing the copy and not writing the tree back to the library.

Changed the code to create a deep copy internally and return a reference
to the caller.  Now if the caller modifies anything in the returned data
structure, the Calset.save method takes care of persisting the change
back to the C code before saving.