Clone the C libvna as c-libvna to avoid name conflict
In the readthedocs build, we're standing in the top-level source directory
of the python package as we work.  When we switched from setuptools
to meson-python, we eliminated the "src" directory, moving the python
"libvna" up where it now conflicts with the C library.  To avoid it,
clone the C library as c-libvna.