Make libvna.cal input array frequency dimension first
Before, the `a` and `b` matrix inputs to the libvna.cal.Solver.add*
functions and the libvna.cal.Calibration.apply function had the dimensions
ordered as (rows x columns x frequencies).  This commit switches the order
to (frequencies x rows x columns), making it consistent with libvna.data,
Touchstone files, and similar software.

Note that this is an incompatible interface change.