Make NPData constructor take filename and filehandle
Enhance the NPData constructor to support loading from a file
or filehandle and doing an in-place conversion if type is given.
This makes it possible to simplify the following sequence:

    npd = NPData()
    npd.load('test.s2p')
    npd = npd.convert(PType.S)

to simply:

    npd = NPData(filename='test.s2p', ptype=PType.S)

It's also more efficient because if the file already contains the desired
parameter type, no copy is needed.