Issue #2: libvna.conv.XtoY functions should accept scalar z0 ...
Modify libvna.conv.canonicalize_z0 so that it always accepts a scalar, a
vector of length number of dimensions in the array, or if an input array
of more than two dimensions is given, an array of shape outer-shape +
(n,) where n is the number of dimensions in the input array.

Add new tests to validate the fix and catch regressions.