MacOS: build libyaml, libvna from source instead of using brew
The MacOS repair wheel step started failing because the tool was changed
to check library compatiblity and was noticing that libyaml and libvna
were both at the version the build machine was running instead of the
target build version.  The `brew install` command does not seem to be
intended to build for older targets.  Abandoned use of brew intall in
favor of building both libraries from source with MACOSX_DEPLOYMENT_TARGET
set appropriately.