Add delay parameters to Solver.add_* and Calibration.apply
This commit adds optional delay parameters to each of the
libvna.cal.Solver.add_* methods and the libvna.cal.Calibration.apply
method.  These parameters are useful when the standard or DUT doesn't
lie exactly on the reference plane and the transmission line is ideal
enough that we can ignore transmission line losses and dispersion.
A positive delay value indicates that the device lies on a piece of
transmission line beyond the reference plane (farther from the VNA); a
negative value indicates that the device lies before the reference plane
(closer to the VNA).  Just as frequency values are expressed in Hz,
delay values are expressed in seconds.