Add reference impedance re-normalization
In libvna.conv, there are three new functions:
    stos(array, z1, z2)
    ttot(array, z1, z2)
    utou(array, z1, z2)

that re-normalize from an original set of reference impedances, z1,
to a new set if references, z2.  These parameters can be either scalar
or a vector with length equal to the number of ports.

Similarly, the stot, stou, ttos, ttou, utos, and utot functions now
take optional z1 and z2 parameters to convert and re-normalize in one
operation.

Finally, in libvna.data, the convert function now takes an optional
new_z0 parameter to convert and re-normalize.  The parameter can be a
scalar, a vector with length equal to number of ports, or a frequencies
by ports array.